+++ CloudShell.ControlPlane/Hosting/CloudShellControlPlaneApplicationBuilderExtensions.cs
@@ -95,7 +95,7 @@ public static class CloudShellControlPlaneApplicationBuilderExtensions
         builder.Services.AddSingleton<IResourceProvider>(
             serviceProvider => serviceProvider.GetRequiredService<LocalHostNetworkProvider>());
         builder.Services.TryAddEnumerable(
-            ServiceDescriptor.Singleton<IResourceProvider, ResourceIdentityProvisioningResourceProvider>());
+            ServiceDescriptor.Scoped<IResourceProvider, ResourceIdentityProvisioningResourceProvider>());
         builder.Services.TryAddEnumerable(
             ServiceDescriptor.Singleton<IResourceProvider, MacOSHostNetworkProvider>());
         builder.Services.TryAddEnumerable(